Splited fhirpath part with aiohttp server definition
Ref #1
